Reverting some changes in commit 6302b98
Check to avoid generation of duplicated stylesheet-tags shouldn't have
been removed from Gridster. It's useful when calling multiple times 
`generate_stylesheet` method once Gridster is initializated. We use it
at ducksboard to generate a TVmode version of the grid with bigger widgets.

Refactored the way stylesheet tags are removed. When created I push them 
to a jQuery collection which is removed when `destroy` method is executed.