Prevent multi <style> generation
If the autogenerate_stylesheet option is set to true then the library is adding multiple times the same style on the head of the document (see the source of the http://gridster.net/demos/adding-widgets-dynamically.html demo in order to observe this). With this correction the gridster style element is only inserted once on the style.