distro script: Replaced outdated method of detecting OpenSUSE version
Replaced the crude parsing of /etc/SuSE-release (which is due to be removed in
future OpenSUSE versions) with equally crude parsing of /etc/os-release.
/etc/os-release is present on several different distributions and usually
contains the same information as /etc/lsb-release when both files are
available.
For now, parsing /etc/lsb-release takes precedence and /etc/os-release is only
used as a fallback when the former is not available.