rrdcached hack to bypass is_file checks
The codebase has many, many is_file checks as part of the polling process even though rrdcached allows creates over the network now.  This hack writes both locally and to the network in order to satisfy the is_file checks.  Please note that the local files will only be created and never updated as updates happen on the network.