Optimize map SQL
Changing LEFT joins to INNER joins brought a 70k result down from 4.1s to 1.7 in my testing.  The only benefit to LEFT joins is to retrieve rows with no match, but we filter those rows out in the WHERE clause.

Was able to map interconnections between 254 devices in 47s with both 'mac' and 'xdp' options turned on.