Revert to existing wifi_clients.inc.php; now LAST.
The first commit of this PR moved to using generic_multiline to graph an
Average of wifi clients and have a legend showing average wifi clients.
This lead to the possiblity of decimal wifi_clients in the legend.

This commit goes the opposite direction.  It retains the original custom
graph (although freshens it up) and changes the graph strategy to LAST.

This means the legend will now match the graph (which it does NOT in the
present code which graphs and average and prints a LAST value).  The
cost is a somewhat more jagged graph since it jumps from integer to
integer value.