Update Centos 7 nginx install steps
Found some issues on a fresh install of Centos 7. PHP defaults to listening on a port, while LibreNMS expects a unix socket. Also, the listen owner has to match the account that nginx runs under.
This is my first proposed change on Github so if I'm doing this all wrong, feel free to smack me down.
Thanks.