append Avaya VSP HW revision to serial number (#6694)
* append Avaya VSP HW revision to serial number

The old way added the HW revision to the sysDescr.  When managing a bunch of switches with different versions this made it look like you had a lot of different kinds of switches.  These are all the switches.  Just add it to the serial number so you can have the hw version for reference if needed.  Just makes more sense not to appended to sysDescr.

* Update avaya-vsp.inc.php

* Update avaya-vsp.inc.php

forgot to add the $hardware variable so Platform would populate

* Update avaya-vsp.inc.php