Set ifIndex to null if there is no dot1dBasePortIfIndex
The ifIndex here is only being used to query for $local_port_id, and when there is no dot1dBasePortIfIndex, the default value is set to the index of the array. This is problematic because the index for interfaces in SNMP for devices do not start from 1. 

ifIndex = 1 is set for VLAN1,
ifIndex = 2 is set for VLAN2, and so forth.....


This results in librenms wrongly attaching the local_port_id to VLan1. 

Setting this to null should not be a problem. The reason for this is that $local_port_id should be totally resolvable from lldpLocPortId.