Provide simple/complex Oxidized configs
This patch provides the ability to override most aspects of the configuration output provided to Oxidized. This should be backwards compatible with any current maps defined previously but allows for extra features such as group overrides in Oxidized as well as IP changes (for hosts that are treated differently inside Oxidized), as well as ProxyHost additions. Of course, this goes further and allows for any flag that can be defined within Oxidized to be mapped in the config.