Update Installation-Ubuntu-1804-Apache.md
The 1604 directions had a mixture of vi and vim for editing files. Since this is ubuntu and we want the instructions to be as easy as possible, I updated editor to be nano. I have no personal preference and can make it whatever we prefer, just figured it should be consistent between the two guides.