Create Installation-Ubuntu-1804-Nginx.md (#8630)
* Create Installation-Ubuntu-1804-Nginx.md

A few tweaks are required to the Ubuntu 16.04 instructions to install on the newly released 18.04 version mostly relating to php being bumped to 7.2 from 7.0 and no mcrypt module for this php version in the distro repository.

* Create Installation-Ubuntu-1804-Apache.md

Used instructions from 1604 Apache as template and updated to work on Bionic (18.04).

* Update Installation-Ubuntu-1804-Nginx.md

Removed instructions for installing mcrypt module as I confirmed it works without it and it was mentioned it was no longer needed.

* Update Installation-Ubuntu-1804-Apache.md

The 1604 directions had a mixture of vi and vim for editing files. Since this is ubuntu and we want the instructions to be as easy as possible, I updated editor to be nano. I have no personal preference and can make it whatever we prefer, just figured it should be consistent between the two guides.

* Update mkdocs.yml

Added Ubuntu 18.04 instruction pages

* Update Installation-Ubuntu-1804-Apache.md

Corrected to libapache2-mod-php7.2 from libapache2-mod-php7.0

* Updated Ubuntu install docs

* Added missing mbstring + curl packages