Extra daily.sh safety If the php update channel check fails, it returns an error code 255... This isn't handled by daily.sh, so it doesn't update code at all.
If updates are disabled (0), skip.
If master (1) or in the php version jail do a git pull update
Otherwise update to the latest tagged release

This should give a "safe" fallback if php isn't working.  Might cause flipping back and forth if the user's error isn't (or can't be) fixed by code changes.