Fix alert follow up for custom queries
RunFollowUp checks if there is a custom sql query attached to
the alert and uses this before invoking GenSQL to build the query.
However, the custom query column was not included when selecting
alerts so the GenSQL function was always invoked.

This patch includes the alert_rules.query in the loadAlerts function
used by RunFollowUp.