Improve UI of filter menu
In the current UI of the filter menu, it's not readily apparent that one, you are supposed to use the filter tabs, and two, that you need to refresh in order for the filter to apply. This request adds button colors to search, update, and refresh to make the filter bar and action buttons stand out.