Fix "fusion" limits detection to unbreak unit test
Commit eda0380 didn't pass the build tests because the behaviour
around limits was changed. As it turns out "fusion" limit detection
was broken all along since YamlDiscovery::getValueFromData() doesn't
like it when the pre-cached value it's trying to find is nested
below a Table OID.

The relevant bit of the array as it was previously pre-cached:

(
    [portTransceiverThresholdInfoTable] => Array
        (
            [25] => Array
                (
                    [portTransceiverThresholdInfoTxBiasCurrentLowAlarm] => 0
                    [portTransceiverThresholdInfoTxBiasCurrentLowWarn] => 10
                    [portTransceiverThresholdInfoTxBiasCurrentHighWarn] => 9000
                    [portTransceiverThresholdInfoTxBiasCurrentHighAlarm] => 10000
                )

        )

)

YamlDiscovery::getValueFromData() would look for the low_limit
OID portTransceiverThresholdInfoTxBiasCurrentLowAlarm in the
pre-cached data, but wouldn't find it under the pre-cached OID
portTransceiverThresholdInfoTable and therefore returns null.

Plot twist: the vendor takes, for example with the dBm graphs, a
value directly expressed in dBm for portOpticalMonitoringInfoRxPower,
and then outputs this as a value that doesn't need a divisor:

1.3.6.1.4.1.259.6.10.120.1.2.11.1.6.28|4|-5.18 dBm normal

But the vendor displays the thresholds in integer32 format with
a divisor of 100 to be applied to get the real value, take let's
say the portTransceiverThresholdInfoRxPowerLowAlarm OID:

1.3.6.1.4.1.259.6.10.120.1.2.12.1.18.28|2|-1640

So, with the same workaround that I built for schleifenbauer I could
also add a different limit divisor for the "fusion" operating system.