preg_match() before calling expLang->evaluate()
The regular expression /^-?\d+(\.\d+)?( ?[-+*\/] ?-?\d+(\.\d+)?)+$/
only matches formulas which begin and end with rational numbers,
and where each rational number is separated by one simple arithmetic
operator. This means the ExpressionLanguage->evaluate() call will not
be made if the variables were not replaced, or if someone tries to
do something fishy.