Support master/slave database operation
Per https://laravel.com/docs/5.1/database#read-write-connections,
configure seperate read/write connections for the database
configuration, so that reads prefer the slave, if listed.  This improves
performance if you have a globally distributed deployment, such as a
master database and GUI installation in North America, and also
follow-the-sun NOCs in Europe and Asia, with their own GUI and slave
database servers.  Default to using the write database as the read if
not configured in .env