Fix LibreNMS Logo SVG missing i sometimes (#10632)
* Update librenms_logo_light.svg

Corrected an issue with the "i" being invisible in certain scenarios on certain (Chromium version 77) browsers, see https://community.librenms.org/t/librenms-svg-logo-issue-in-some-zooms/9640/4. The issue seems to be that "i" was a separate clip-path from the "L bre".

* Update librenms_logo_blue.svg

Corrected an issue with the "i" being invisible in certain scenarios on certain (Chromium version 77) browsers, see https://community.librenms.org/t/librenms-svg-logo-issue-in-some-zooms/9640/4. The issue seems to be that "i" was a separate clip-path from the "L bre".

* Update librenms_logo_mono.svg

Corrected an issue with the "i" being invisible in certain scenarios on certain (Chromium version 77) browsers, see https://community.librenms.org/t/librenms-svg-logo-issue-in-some-zooms/9640/4. The issue seems to be that "i" was a separate clip-path from the "L bre".

* Update librenms_logo_dark.svg

Corrected an issue with the "i" being invisible in certain scenarios on certain (Chromium version 77) browsers, see https://community.librenms.org/t/librenms-svg-logo-issue-in-some-zooms/9640/4. The issue seems to be that "i" was a separate clip-path from the "L bre".