Fixed syslog pruning when dbFetchRow() returns array (#10850)
If dbFetchRow() returns an array, dbDelete() won't do its work pruning the syslog table and we'll drop out of the block via break.  This if statement will take the value return inside the array and replace $limit with that value.  This way, dbDelete() can successfully prune the syslog table.

dbFetchCell() can return an array that won't work when fed to dbDelete().  This keeps that from occurring so the syslog table can be pruned.