Prevent syslog table purge from spamming daily.log (#10851)
daily.php's syslog purge block writes output for every dbDelete() except for the final one.  The deletes are done in blocks of 1000 rows.  This can lead to verbose output while the table is purged.  This patch saves up all that excitement for a single echo at the end of the operation.  It also outputs the total amount of rows deleted.