Fix bills api output for period=previous
Include static fields from the bills table in the previous period output.
Check lowercase(bill_type) because the history table uses CDR and Quota.
Update SQL IF() because the bills table always uses lowercase.

Because of these changes the calculation of the fields allowed, used,
percent and overuse for period=previous is now enabled.