Clear if/mount/sensor filters before generating test data
If you're generating test data from a real device, you're probably
doing so on a production instance (or at least, that's where I
poll all of my most expensive kit from!).

This means there are likely to be a bunch of config settings that
cause test data to be generated differently between CI and your
own instance.

This change clears the if, mount and sensor filters before test
data is retrieved from SNMPSim.