Create CentOS-8-Apache installation
This method covers installation of LibreNMS with Apache on CentOS 8, with Remi Repo and Dispatcher Service as default. It has been tested 4 times with new installs, all 4 times were successful. This document is by no means as complete as the CentOS7 one, but I would rather get a basic working install in the docs, than an elaborate one in 3 months time. If you guys are not happy with this install, please delete it, I would rather not create any confusion.