Make IPv6 discovery work on JunOS
Although JunOS populates ipAddressIfIndex.ipv6, it does not add prefix lengths or other metadata. This means that IPv6 discovery fails.

This change removes OIDs from the list when there is incomplete data. If the list is empty after IP-MIB is walked, it then tries IPv6 MIB.

It's not a great fix, but I'm not confident I can fix the underlying issue without a complete re-write of this module.