Update hpe-ilo.yaml
Added support for Logical Drive Status.

When using 3rd party disks with for example P420i, the controller reports the Logical Drive Condition as degraded as the disk is unauthenticated. So to get the actual status of the logical device one has to look at Logical Drive Status instead (which also reports a lot of other states).