Update dell-compellent.yaml
Dell changed the text for sysDescr after the 2017 firmware-release so that the string "COMPELLEN" which was looked after here before, did not appear anymore. The Compellent cuts off the snmp-answer after 255 characters.
I replaced the previous detection to a new snmpget-detection using the oid (DELL-STORAGE-SC-MIB::productIDDisplayName.0) that shouldnt be altered in the future.