Improve poller and wrapper effiencey
Reduce the CPU usage of the polling process by getting the wrapper to run a persistent copy of poller.php for each thread.

This commit also modifies the queueuing for the poll process to try and have a more even use of the CPU throughout the entire poll window.  This is achieved by:
 - Splitting the device list in half
 - Splitting the workers across the 2 queues to assign more workers to the devices that take longer to poll
 - Making the number of workers requested on the command line the maximum, but choosing a number of workers that will complete just before the next group will start.