Adjust low and high default temperature thresholds
TIA-942 states a datacentre should have a minumum temperature of 18 degrees. This works as a good value for both chassis cooling sensors, and for environmental as it's slightly around the point where condensation can start to form in humid environments.

Maximum temperature is a little trickier, as it varies depending on context. For an environmental sensor in a data centre, 27 degrees is the bad number. In a wiring cupboard though, it could be 40 degrees or so.

A quick sample of Intel CPU specifications list the maximum temperature above 70 for core temperature, up to 100. I can't find any hard numbers but it looks like Turbo enabled processors start throttling around 80 degrees. Checking my own instance, around 3% of the temperature sensors are exceed that threshold; all under heavy load.

The real solution is for thresholds to be provided by the vendor, but one of the most common sensor types (lm_sensors) does not do this.