Change text to show alert termplate msg
In the initial pull request this was brought up but never addressed, https://github.com/librenms/librenms/pull/9305#issuecomment-427930650  The only reason i can think of is due to SMS message length, but you could be mindful of that with a alert template.  Submitting this to see if anyone objects to changing it.