Add battery operating time for APC USVs
This time is calculated as the difference between current day
and the day of the last battery replacement. It is stored in
minutes. In the GUI the value will be display as years and days.

A value of at least 4 years will be treated as 'yellow' level (warning).
A value of at least 5 yearswill be treated as 'red' level (critical).
These are just our personal preferences.