Improve behaviour of ssh/web/telnet buttons.
As of now, these 3 buttons (located beyond the 3 dot panel at the almost-top-right of the device page)
will use only the hostname ... which is bad, if you do not want to access your devices
via hostname, but instead via ip.

Therefore, you can now use the config setting device_external_link.use_assigned_ip to ensure,
that your links a generated with the overwrite IP (this one must be set, of course):

    $config['device_external_link']['use_assigned_ip'] = true;

You may also enable the usage of the system name (sysName),
if this one contains your FQN for whatever reason:

    $config['device_external_link']['use_sysName'] = true;

Otherwise, the current behaviour is kept – using the hostname for building the links.

Along the way, the title value has also been adjusted slightly … you know, for good measure and so on …