API: update_device tweaks
Changes the update_device function to simplify it, and allow setting only location. API allows either a single field or an array of fields. Location could not be set unless it was sent as array. Combine the two paths into one by converting the single field case to an array and then doing the regular work on an array, with a small check on the result to send the old single field output message for compatibility.