feat: Add support for filtering by device groups
This feature works for polling, discovery and also services
You may either use the id of a device group or it’s name; using a wildcard
instead of a name (e.g. MyGr* or *Group instead of MyGroup) is possible, too.

Please note: Using a name (or name part) containing a whitespace or a comma will not work as expected; escaping by quotes etc. is not implemented

Motivation: Some modules shall be run more often than others, e. g. discovering the bgp sessions on a Nexus router. Therefore you want to be able to filter by a group (e.g. “Nexus_Devices”)