Add support for FiberStore "data center" switches
FiberStore has two different model lines of switches, and this adds
support for the "data center" line.  This was configured and tested
against the N5860-48SC model.

The MIBs are from FiberStore's website, with the exception of the
FS-FIBER-MIB (the website version is a little behind the software, so
support sent me an updated version and should be getting it published
soon).

This does not currently handle 4x10 or 4x25 QSFP+ light levels, as I
don't currently have any to test against.