Add support for Nokia 1830 PSS Optical gear
Tested against 1830 PSS-8 Chassis, in single and multi-chassis configurations
Supports Optical Power, temperature, current, power, and a variety of card states, alarms, and trail status.

Of note: With multi-chassis nodes, reading values from the non-master can be a "slow" process, slow in that the default timeout Libre ships with can cause the SNMP client to begin issuing retries.  This can result in the chassis daemon rate limiting and/or blocking the client.  The solution here is to raise the SNMP timeout within each device's configuration; unfortunately those values are not configurable on a per-OS basis (at least not at this time).