Modified config options for mysql database to make SSL/TLS settings optional.
SSL/TLS CA, CERT, and KEY are optional. If they are not provided the array_filter() will remove them.

MYSQL_ATTR_SSL_VERIFY_SERVER_CERT is outside array_filter() because the function will remove the option if it is falsey; however there are times when that variable must be set to false.