Add nfsen_base to config_definitions.json (#16065)
* Update config_definitions.json

Added nfsen_base so that nfsen stats feature works.  I had to add as an array as it was the way I was able to make it work, even though another type may make more sense from a user interface perspective.  I put it in order 10 as there seemed to be a gap there, but one could argue that it should be higher and other nfsen items renumbered.

* Update config_definitions.json

Fixed typo

* Update config_definitions.json

* Update settings.php

Added Default NFSen Base Directory

* Update settings.php

---------

Co-authored-by: Tony Murray <murraytony@gmail.com>