LLDP Discovery - LldpRemPortId convert to string when in HEX (#16438)
* Update discovery-protocols.inc.php

When using the LLDP-V2-MIB, the device may reply with lldpV2RemPortId as hexadecimal. To convert it to a string, this update checks if the lldpV2RemPortIdSubtype is '5' = "interfaceName(5)"; if so, the code removes the character ":" from the value of lldpV2RemPortId and tests if it is hexadecimal. If the test is positive, it translates it into a string.

* Update discovery-protocols.inc.php
Added lines 342 to 345 to convert the lldpRemPortId when is hex and lldpRemPortIdSutype is 5.

* Update discovery-protocols.inc.php
Importing lib StringHelpers.

* tabs + tests PANOS

---------

Co-authored-by: PipoCanaja <38363551+PipoCanaja@users.noreply.github.com>