When mail sending fails, throw an AlertTransportDeliveryException.
Currently, when mailing with PHPMailer fails, the system seems to
think it was successful. This reworks the Mail::send method a bit to
throw exceptions instead of returning a string when delivery fails. In
the Mail transport, the exceptions are translated into an
AlertTransportDeliveryException. This makes it so that when you do
"Test Transport" on an email alert transport, it will show the failure
instead of saying it was successful.