convert enums in mpls_sdp_binds to varchars
There was a typo in the enum sdpBindTypes of the table mpls_sdp_binds.
The word "mirror" was mispelled as "mirrior" and "atmSdu" and "atmCell"
were concatinated for some reason, which breaks pollers and discovery.
This commit changes the column-type of all enum columns of this table
to simple strings to prevent such issues in the future.