Fix for Generex UPS firmware 2.30 detection
Somewhere between firmware 2.16 and 2.30 Generex has changed the output of sysDescr. It used to contain the model string (eg CS141). That seems to have changed to '<oem> SNMP Agent', where <oem> can be a range of brands.

See https://www.generex.de/support/downloads/ups/cs141/update for a list of brands that use this card