Make snmpwalk_cache_oid a wrapper around new style SnmpQuery
The original snmpwalk_cache_oid does stripping/trimming, causing some
strings to be misrepresented. SnmpQuery does handle things properly.

Codebase should be updated to use SnmpQuery, but by adding this wrapper,
we get the benefits of SnmpQuery for a multitude of original calls
immedaitely.

In this particular case it fixes an InvalidIpException thrown when an IP
address happened to contain only readable ascii and start (or end) with
a backslash.

Before this patch, the IP 92.65.66.67 would give an "ABC" parse error,
while it should have been "\\ABC", which is readable.

This patch can be amended to handle more cases and then
snmpwalk_cache_oid_orig can be phased out sooner.

Issue: #16992