Fix Proxmox module: Correct parameter usage for VM and device ID (#17114)
* Fix Proxmox module: Use $vars array for VM info retrieval

Replaced Request::get() with $vars[] in proxmox_vm_info() to properly retrieve VM ID and instance.
This fixes an issue where the Proxmox module was not working due to incorrect parameter retrieval.

* Fix Proxmox module: Use correct device ID for VM retrieval

Replaced Request::get('device') with $device["device_id"] in proxmox_node_vms() to ensure the correct device ID is passed.
This fixes an issue where the Proxmox module was not retrieving VMs correctly.

* Update proxmox.inc.php

---------

Co-authored-by: Neil Lathwood <gh+n@laf.io>