Fix header mode authentication in SSOAuthorizer
When this function is called: authSSOGetAttr($attr, $prefix = 'HTTP_')
$prefix is alway set to '' due to this call: return $this->authSSOGetAttr(Config::get('sso.user_attr'), '');

This breaks the header matching from Nginx which is prefixed with HTTP_. This fix allows the default of $prefix = 'HTTP_' to be used.

This fixes Authelia using the following settings:
    "user_attr": "REMOTE_USER",
    "mode": "header",
    "group_strategy": "static",
    "static_level": 10

Tested with docker image librenms/librenms:25.2.0, using Librenms helm chart version 3.18.0.