Get rid of dashboard white tint (#18507)
* Get rid of annoying dashboard white tint

Everything on the dashboard looks whiter than it should. This is especially noticeable when using a map with satellite view.

Based on the spacing above/below the opacity line, this seems to be a debugging thing from 10 years ago when the dashboard was initially created.

* Clean up CSS transitions in .gridster class

Removed vendor prefixes for CSS transitions in .gridster class.

---------

Co-authored-by: Tony Murray <murraytony@gmail.com>